+++ web-console/src/lib/components/connectors/dialogs/SnowflakeOutputConnector.tsx
@@ -121,12 +121,7 @@ export const SnowflakeOutputConnectorDialog = (props: ConnectorDialogProps) => {
 
   const tabList = ['detailsTab', 'sourceTab', 'authTab', 'formatTab']
   const tabFooter = (
-    <TabFooter
-      isUpdate={props.connector !== undefined}
-      activeTab={activeTab}
-      setActiveTab={setActiveTab}
-      tabsArr={tabList}
-    />
+    <TabFooter submitButton={props.submitButton} activeTab={activeTab} setActiveTab={setActiveTab} tabsArr={tabList} />
   )
 
   return (
@@ -159,7 +154,9 @@ export const SnowflakeOutputConnectorDialog = (props: ConnectorDialogProps) => {
           </IconButton>
           <Box sx={{ mb: 8, textAlign: 'center' }}>
             <Typography variant='h5' sx={{ mb: 3 }}>
-              {props.connector === undefined ? 'New Snowflake-Kafka Output' : 'Update ' + props.connector.name}
+              {props.connector === undefined
+                ? 'New Snowflake-Kafka Output'
+                : props.existingTitle?.(props.connector.name) ?? ''}
             </Typography>
             {props.connector === undefined && (
               <Typography variant='body2'>Output to a Snowflake table via a Kafka topic</Typography>
@@ -236,25 +233,25 @@ export const SnowflakeOutputConnectorDialog = (props: ConnectorDialogProps) => {
                 value='detailsTab'
                 sx={{ border: 0, boxShadow: 0, width: '100%', backgroundColor: 'transparent' }}
               >
-                <TabKafkaNameAndDesc direction={Direction.OUTPUT} />
+                <TabKafkaNameAndDesc direction={Direction.OUTPUT} disabled={props.disabled} />
                 {tabFooter}
               </TabPanel>
               <TabPanel
                 value='sourceTab'
                 sx={{ border: 0, boxShadow: 0, width: '100%', backgroundColor: 'transparent' }}
               >
-                <TabkafkaOutputDetails />
+                <TabkafkaOutputDetails disabled={props.disabled} />
                 {tabFooter}
               </TabPanel>
               <TabPanel value='authTab' sx={{ border: 0, boxShadow: 0, width: '100%', backgroundColor: 'transparent' }}>
-                <TabKafkaAuth />
+                <TabKafkaAuth disabled={props.disabled} />
                 {tabFooter}
               </TabPanel>
               <TabPanel
                 value='formatTab'
                 sx={{ border: 0, boxShadow: 0, width: '100%', backgroundColor: 'transparent' }}
               >
-                <SnowflakeOutputFormatDetails />
+                <SnowflakeOutputFormatDetails disabled={props.disabled} />
                 {tabFooter}
               </TabPanel>
             </TabContext>